LPS-176408 Create a base to avoid duplications
