+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/info/collection/provider/ManyToManyObjectRelationshipRelatedInfoCollectionProvider.java
@@ -14,8 +14,6 @@
 
 package com.liferay.object.internal.info.collection.provider;
 
-import com.liferay.info.collection.provider.CollectionQuery;
-import com.liferay.info.collection.provider.RelatedInfoItemCollectionProvider;
 import com.liferay.info.pagination.InfoPage;
 import com.liferay.info.pagination.Pagination;
 import com.liferay.object.model.ObjectDefinition;
@@ -23,21 +21,13 @@ import com.liferay.object.model.ObjectEntry;
 import com.liferay.object.model.ObjectRelationship;
 import com.liferay.object.service.ObjectDefinitionLocalService;
 import com.liferay.object.service.ObjectEntryLocalService;
-import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
-
-import java.util.Collections;
-import java.util.Locale;
 
 /**
  * @author Jürgen Kappler
  */
 public class ManyToManyObjectRelationshipRelatedInfoCollectionProvider
-	implements RelatedInfoItemCollectionProvider {
+	extends BaseObjectRelationshipRelatedInfoCollectionProvider {
 
 	public ManyToManyObjectRelationshipRelatedInfoCollectionProvider(
 			ObjectDefinition objectDefinition,
@@ -46,94 +36,29 @@ public class ManyToManyObjectRelationshipRelatedInfoCollectionProvider
 			ObjectEntryLocalService objectEntryLocalService)
 		throws PortalException {
 
-		_objectDefinition = objectDefinition;
-		_objectRelationship = objectRelationship;
-		_objectEntryLocalService = objectEntryLocalService;
-
-		_relatedObjectDefinition =
-			objectDefinitionLocalService.getObjectDefinition(
-				_objectRelationship.getObjectDefinitionId2());
+		super(
+			objectDefinition, objectDefinitionLocalService,
+			objectEntryLocalService, objectRelationship);
 	}
 
 	@Override
-	public InfoPage<ObjectEntry> getCollectionInfoPage(
-		CollectionQuery collectionQuery) {
-
-		Object relatedItem = collectionQuery.getRelatedItem();
-
-		if (!(relatedItem instanceof ObjectEntry)) {
-			return InfoPage.of(
-				Collections.emptyList(), collectionQuery.getPagination(), 0);
-		}
-
-		ObjectEntry objectEntry = (ObjectEntry)relatedItem;
-
-		Pagination pagination = collectionQuery.getPagination();
-
-		try {
-			return InfoPage.of(
-				_objectEntryLocalService.getManyToManyObjectEntries(
-					objectEntry.getGroupId(),
-					_objectRelationship.getObjectRelationshipId(),
-					objectEntry.getObjectEntryId(), true,
-					_objectRelationship.isReverse(), pagination.getStart(),
-					pagination.getEnd()),
-				collectionQuery.getPagination(),
-				_objectEntryLocalService.getManyToManyObjectEntriesCount(
-					objectEntry.getGroupId(),
-					_objectRelationship.getObjectRelationshipId(),
-					objectEntry.getObjectEntryId(), true,
-					_objectRelationship.isReverse()));
-		}
-		catch (PortalException portalException) {
-			_log.error(portalException);
-		}
+	protected InfoPage<ObjectEntry> getCollectionInfoPage(
+			ObjectEntry objectEntry, Pagination pagination)
+		throws PortalException {
 
 		return InfoPage.of(
-			Collections.emptyList(), collectionQuery.getPagination(), 0);
+			objectEntryLocalService.getManyToManyObjectEntries(
+				objectEntry.getGroupId(),
+				objectRelationship.getObjectRelationshipId(),
+				objectEntry.getObjectEntryId(), true,
+				objectRelationship.isReverse(), pagination.getStart(),
+				pagination.getEnd()),
+			pagination,
+			objectEntryLocalService.getManyToManyObjectEntriesCount(
+				objectEntry.getGroupId(),
+				objectRelationship.getObjectRelationshipId(),
+				objectEntry.getObjectEntryId(), true,
+				objectRelationship.isReverse()));
 	}
-
-	@Override
-	public String getCollectionItemClassName() {
-		return _relatedObjectDefinition.getClassName();
-	}
-
-	@Override
-	public String getKey() {
-		return StringBundler.concat(
-			RelatedInfoItemCollectionProvider.super.getKey(), "_",
-			_objectDefinition.getCompanyId(), "_", _objectDefinition.getName(),
-			"_", _objectRelationship.getName());
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return _relatedObjectDefinition.getPluralLabel(locale);
-	}
-
-	@Override
-	public String getSourceItemClassName() {
-		return _objectDefinition.getClassName();
-	}
-
-	@Override
-	public boolean isAvailable() {
-		if (!FeatureFlagManagerUtil.isEnabled("LPS-176083") ||
-			(_objectDefinition.getCompanyId() !=
-				CompanyThreadLocal.getCompanyId())) {
-
-			return false;
-		}
-
-		return true;
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		ManyToManyObjectRelationshipRelatedInfoCollectionProvider.class);
-
-	private final ObjectDefinition _objectDefinition;
-	private final ObjectEntryLocalService _objectEntryLocalService;
-	private final ObjectRelationship _objectRelationship;
-	private final ObjectDefinition _relatedObjectDefinition;
 
 }
